resolve merge conflict

+++ classes/class.ilObjInteractiveVideo.php
@@ -399,14 +399,15 @@ class ilObjInteractiveVideo extends ilObjectPlugin implements ilLPStatusPluginIn
 	{
         if (((!$this->referenced) || ($this->countReferences() == 1)) && $this->video_source_object !== null ) {
             $this->getVideoSourceObject($this->getSourceId());
-            if($this->video_source_object !== null) {
-                $this->video_source_object->beforeDeleteVideoSource($this->getId());
+            if($this->getId() !== null) {
+                if($this->video_source_object !== null) {
+                    $this->video_source_object->beforeDeleteVideoSource($this->getId());
+                }
+                self::deleteComments(self::getCommentIdsByObjId($this->getId(), false));
+
+                $this->db->manipulate('DELETE FROM ' . self::TABLE_NAME_OBJECTS . ' WHERE obj_id = ' . $this->db->quote($this->getId(), 'integer'));
+                $this->deleteMetaData();
             }
-
-            self::deleteComments(self::getCommentIdsByObjId($this->getId(), false));
-
-            $this->db->manipulate('DELETE FROM ' . self::TABLE_NAME_OBJECTS . ' WHERE obj_id = ' . $this->db->quote($this->getId(), 'integer'));
-            $this->deleteMetaData();
         }
         return true;
 	}
